+++ package.json
@@ -6,6 +6,8 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
+    "dotenv": "^16.0.0",
+    "env": "0.0.2",
     "multiselect-react-dropdown": "^2.0.17",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
