+++ src/Components/Recipes/recipes.js
@@ -4,7 +4,7 @@ const Recipes = () => {
   const { query, number, intolerences } = JSON.parse(
     localStorage.getItem("filter")
   );
-  const url = `https://api.spoonacular.com/food/ingredients/search?apiKey=f901b8871ffc4194845cf8e20a40b7b0&query=${query}&intolerances=${intolerences}&number=${number}`;
+  const url = `https://api.spoonacular.com/food/ingredients/search?apiKey=${process.env.REACT_APP_API_KEY}&query=${query}&intolerances=${intolerences}&number=${number}`;
   const { products, loading } = useFetch(url);
   const { results } = products;
 
@@ -26,7 +26,6 @@ const Recipes = () => {
           return (
             <div className="item" key={id}>
               <h4>{name}</h4>
-              <p>{image}</p>
             </div>
           );
         })
