+++ src/Protected-Routes/protectedRoute.js
@@ -3,8 +3,8 @@ import React from "react";
 import { Outlet } from "react-router-dom";
 import Login from "../Components/Login/login";
 
-const ProtectedRoute = ({ isAuth, loading, products }) => {
-  isAuth = localStorage.getItem("logged-in-user") === null ? false : true;
+const ProtectedRoute = ({ loading, products }) => {
+  const isAuth = localStorage.getItem("logged-in-user") ? false : true;
   return isAuth ? <Outlet /> : <Login loading={loading} products={products} />;
 };
